database/requirements.txt
# Database requirements
psycopg2-binary==2.9.6